+++ src/AndreyAkaSkif.ServiceResult.AspNetCore/ErrorMessage.cs
@@ -3,19 +3,12 @@ namespace AndreyAkaSkif.ServiceResult.AspNetCore;
 /// <summary>
 /// Обёртка для тела HTTP-ответа с сообщением об ошибке
 /// </summary>
-public readonly struct ErrorMessage
+/// <param name="title">Сообщение об ошибке</param>
+public readonly struct ErrorMessage(string? title)
 {
-    /// <summary>
-    /// Конструктор
-    /// </summary>
-    /// <param name="title">Сообщение об ошибке</param>
-    public ErrorMessage(string? title)
-    {
-        Title = title;
-    }
 
     /// <summary>
     /// Сообщение об ошибке
     /// </summary>
-    public string? Title { get; }
+    public string? Title { get; } = title;
 }