+++ src/AndreyAkaSkif.ServiceResult/Abstractions/Result.cs
@@ -9,7 +9,7 @@ public abstract class Result<T>
     /// <summary>
     /// Результат успешен
     /// </summary>
-    public bool IsOk { get; protected set; } = false;
+    public bool IsOk { get; protected set; }
 
     /// <summary>
     /// Резельтат неуспешен
@@ -19,12 +19,12 @@ public abstract class Result<T>
     /// <summary>
     /// Сообщение об ошибке
     /// </summary>
-    public string? Error { get; protected set; } = default;
+    public string? Error { get; protected set; }
 
     /// <summary>
     /// Возвращаемый ресурс
     /// </summary>
-    public T? Data { get; protected set; } = default;
+    public T? Data { get; protected set; }
 }
 
 
@@ -36,7 +36,7 @@ public abstract class Result
     /// <summary>
     /// Результат успешен
     /// </summary>
-    public bool IsOk { get; protected set; } = false;
+    public bool IsOk { get; protected set; }
 
     /// <summary>
     /// Резельтат неуспешен
@@ -46,5 +46,5 @@ public abstract class Result
     /// <summary>
     /// Сообщение об ошибке
     /// </summary>
-    public string? Error { get; protected set; } = default;
+    public string? Error { get; protected set; }
 }